requirements.txt
Django>=4.2,<5.0
reportlab==3.6.13
pdfrw==0.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.2.0
stripe==8.0.0
whitenoise==6.6.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pymupdf>=1.22
unidecode==1.3.8
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
numpy<1.25
